+++ src/main/resources/reference.conf
@@ -1,4 +1,4 @@
-play.modules.enabled += dk.jyllandsposten.playprometheusfilters.PrometheusModule
+play.modules.enabled += io.github.jyllandsposten.playprometheusfilters.PrometheusModule
 
 play-prometheus-filters {
   # Registers some default collectors for jvm metrics. See DefaultExports class.
