--- src/main/scala/dk/jyllandsposten/playprometheusfilters/filters/StatusAndRouteLatencyFilter.scala
@@ -1,22 +1,0 @@
-package dk.jyllandsposten.playprometheusfilters.filters
-
-import akka.stream.Materializer
-import dk.jyllandsposten.playprometheusfilters.metrics.DefaultPlayUnmatchedDefaults
-import dk.jyllandsposten.playprometheusfilters.metrics.LatencyRequestMetrics.LatencyRequestMetricsBuilder
-import io.prometheus.client.CollectorRegistry
-import javax.inject.{Inject, Singleton}
-import play.api.Configuration
-
-import scala.concurrent.ExecutionContext
-
-/**
-  * A [[MetricsFilter]] using a histogram metric to record latency.
-  * Latency metric adds 'RouteActionMethod', 'Status', 'Controller', 'Path' and 'Verb' labels.
-  */
-@Singleton
-class StatusAndRouteLatencyFilter @Inject()(registry: CollectorRegistry, configuration: Configuration)(implicit mat: Materializer, ec: ExecutionContext) extends MetricsFilter(configuration) {
-
-  override val metrics = List(
-    LatencyRequestMetricsBuilder.build(registry, DefaultPlayUnmatchedDefaults)
-  )
-}
