--- src/main/scala/dk/jyllandsposten/playprometheusfilters/utils/WriterAdapter.scala
@@ -1,14 +1,0 @@
-package dk.jyllandsposten.playprometheusfilters.utils
-
-import java.io.Writer
-
-class WriterAdapter(buffer: StringBuilder) extends Writer {
-
-  override def write(charArray: Array[Char], offset: Int, length: Int): Unit = {
-    buffer ++= new String(new String(charArray, offset, length).getBytes("UTF-8"), "UTF-8")
-  }
-
-  override def flush(): Unit = {}
-
-  override def close(): Unit = {}
-}
