+++ src/main/scala/io/github/jyllandsposten/playprometheusfilters/filters/StatusAndRouteLatencyAndCounterFilter.scala
@@ -1,0 +1,26 @@
+package io.github.jyllandsposten.playprometheusfilters.filters
+
+import akka.stream.Materializer
+import io.github.jyllandsposten.playprometheusfilters.metrics.CounterRequestMetrics.CounterRequestMetricBuilder
+import io.github.jyllandsposten.playprometheusfilters.metrics.DefaultPlayUnmatchedDefaults
+import io.github.jyllandsposten.playprometheusfilters.metrics.LatencyRequestMetrics.LatencyRequestMetricsBuilder
+import io.prometheus.client._
+import javax.inject.{Inject, Singleton}
+import play.api.Configuration
+
+import scala.concurrent.ExecutionContext
+
+/**
+  * A [[MetricsFilter]] using a both a histogram and counter metric to record latency and count requests.
+  *
+  * Latency metric adds 'RouteActionMethod', 'Status', 'Controller', 'Path' and 'Verb' labels.  
+  * Counter metric adds 'method', 'status', 'controller', 'path' and 'verb' labels.
+  */
+@Singleton
+class StatusAndRouteLatencyAndCounterFilter @Inject()(registry: CollectorRegistry, configuration: Configuration)(implicit mat: Materializer, ec: ExecutionContext) extends MetricsFilter(configuration) {
+
+  override val metrics = List(
+    LatencyRequestMetricsBuilder.build(registry, DefaultPlayUnmatchedDefaults),
+    CounterRequestMetricBuilder.build(registry, DefaultPlayUnmatchedDefaults)
+  )
+}
