+++ src/main/scala/io/github/jyllandsposten/playprometheusfilters/filters/StatusCounterFilter.scala
@@ -1,0 +1,22 @@
+package io.github.jyllandsposten.playprometheusfilters.filters
+
+import akka.stream.Materializer
+import io.github.jyllandsposten.playprometheusfilters.metrics.CounterRequestMetrics.StatusCounterRequestMetricBuilder
+import io.github.jyllandsposten.playprometheusfilters.metrics.DefaultPlayUnmatchedDefaults
+import io.prometheus.client.CollectorRegistry
+import javax.inject.{Inject, Singleton}
+import play.api.Configuration
+
+import scala.concurrent.ExecutionContext
+
+/**
+  * A [[MetricsFilter]] using a counter metric to count requests statuses.
+  * Only adds a 'status' label containing the status codes.
+  */
+@Singleton
+class StatusCounterFilter @Inject()(registry: CollectorRegistry, configuration: Configuration)(implicit mat: Materializer, ec: ExecutionContext) extends MetricsFilter(configuration) {
+
+  override val metrics = List(
+    StatusCounterRequestMetricBuilder.build(registry, DefaultPlayUnmatchedDefaults)
+  )
+}
