+++ src/test/scala/io/github/jyllandsposten/playprometheusfilters/filters/StatusCounterFilterSpec.scala
@@ -1,0 +1,47 @@
+package io.github.jyllandsposten.playprometheusfilters.filters
+
+import io.github.jyllandsposten.playprometheusfilters.mocks.MockController
+import io.prometheus.client.CollectorRegistry
+import org.mockito.ArgumentMatchers.any
+import org.mockito.Mockito.verify
+import org.scalatestplus.mockito.MockitoSugar
+import org.scalatest.{MustMatchers, WordSpec}
+import org.scalatestplus.play.guice.GuiceOneAppPerSuite
+import play.api.Configuration
+import play.api.mvc.Results
+import play.api.test.Helpers.stubControllerComponents
+import play.api.test.{DefaultAwaitTimeout, FakeRequest, FutureAwaits}
+
+import scala.concurrent.ExecutionContext.Implicits.global
+
+class StatusCounterFilterSpec extends WordSpec with MustMatchers with MockitoSugar with Results with DefaultAwaitTimeout with FutureAwaits with GuiceOneAppPerSuite {
+
+  private implicit val mat = app.materializer
+  private val configuration = mock[Configuration]
+
+  "Filter constructor" should {
+    "Add a counter to the prometheus registry" in {
+      val collectorRegistry = mock[CollectorRegistry]
+      new StatusCounterFilter(collectorRegistry, configuration)
+      verify(collectorRegistry).register(any())
+    }
+  }
+
+  "Apply method" should {
+    "Count the requests with status" in {
+      val filter = new StatusCounterFilter(mock[CollectorRegistry], configuration)
+      val rh = FakeRequest()
+      val action = new MockController(stubControllerComponents()).ok
+
+      await(filter(action)(rh).run())
+
+      val metrics = filter.metrics(0).metric.collect()
+      metrics must have size 1
+      val samples = metrics.get(0).samples
+      samples.get(0).value mustBe 1.0
+      samples.get(0).labelValues must have size 1
+      samples.get(0).labelValues.get(0) mustBe "200"
+    }
+  }
+
+}
