+++ src/test/scala/io/github/jyllandsposten/playprometheusfilters/mocks/MockController.scala
@@ -1,0 +1,17 @@
+package io.github.jyllandsposten.playprometheusfilters.mocks
+
+import javax.inject.Inject
+
+import play.api.mvc.{AbstractController, ControllerComponents}
+
+class MockController @Inject()(cc: ControllerComponents) extends AbstractController(cc) {
+
+  def ok = Action {
+    Ok("ok")
+  }
+
+  def error = Action {
+    NotFound("error")
+  }
+
+}
